Refactor purchase feature: migrate purchase-related files to the new feature directory and remove obsolete code
