+++ lib/src/core/router/app_router.dart
@@ -12,7 +12,7 @@ import 'package:quickdrop/src/presentation/financial_information/financial_infor
 import 'package:quickdrop/src/presentation/loading/loading.dart';
 import 'package:quickdrop/src/features/auth/presentation/login/login.dart';
 import 'package:quickdrop/src/presentation/my_purchases/mypurchases.dart';
-import 'package:quickdrop/src/presentation/pruchase/purchase.dart';
+import 'package:quickdrop/src/features/pruchase/presentation/purchase.dart';
 import 'package:quickdrop/src/presentation/purchase_detail/purchase_detail.dart';
 import 'package:quickdrop/src/features/search/search.dart';
