+++ lib/src/features/productDetail/presentation/widgets/quantity/quantity.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
-import 'package:quickdrop/src/presentation/pruchase/purchaseCubit/purchase_cubit.dart';
+import 'package:quickdrop/src/features/pruchase/presentation/cubit/purchase_cubit.dart';
 import 'package:quickdrop/src/features/productDetail/presentation/widgets/quantity/widgets/quantity_button.dart';
 import 'package:quickdrop/src/features/productDetail/presentation/widgets/quantity/widgets/quantity_text_field.dart';
