+++ lib/src/features/pruchase/domain/usecase/purshase_usecase.dart
@@ -1,5 +1,5 @@
 import 'package:quickdrop/src/domain/entity/purchase_entity.dart';
-import 'package:quickdrop/src/domain/repository/purchase_repository.dart';
+import 'package:quickdrop/src/features/pruchase/domain/repository/purchase_repository.dart';
 
 class PurchaseUsecase {
   PurchaseUsecase({
