+++ lib/src/features/pruchase/presentation/widgets/confirm_button/widgets/information_tile.dart
