+++ lib/src/features/pruchase/presentation/widgets/payment_method.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 // import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
-import 'package:quickdrop/src/presentation/pruchase/widgets/card_tile.dart';
+import 'package:quickdrop/src/features/pruchase/presentation/widgets/card_tile.dart';
 // import 'package:quickdrop/src/presentation/pruchase/purchaseCubit/purchase_cubit.dart';
-import 'package:quickdrop/src/presentation/pruchase/widgets/payment_section.dart';
+import 'package:quickdrop/src/features/pruchase/presentation/widgets/payment_section.dart';
 
 class PaymentMethod extends StatefulWidget {
   const PaymentMethod({super.key});
