+++ lib/src/features/pruchase/presentation/widgets/payment_section.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:quickdrop/src/core/constants/constants.dart';
-import 'package:quickdrop/src/presentation/pruchase/widgets/title_section.dart';
+import 'package:quickdrop/src/features/pruchase/presentation/widgets/title_section.dart';
 
 class PaymentSection extends StatelessWidget {
   const PaymentSection({
