+++ lib/src/injection/injection_container.dart
@@ -6,7 +6,7 @@ import 'package:quickdrop/src/features/auth/data/datasource/auth_datasource.dart
 import 'package:quickdrop/src/features/home/data/datasource/home_datasource.dart';
 import 'package:quickdrop/src/data/datasource/my_purchases_datasource.dart';
 import 'package:quickdrop/src/features/productDetail/data/datasource/produt_datasource.dart';
-import 'package:quickdrop/src/data/datasource/purchase_datasource.dart';
+import 'package:quickdrop/src/features/pruchase/data/datasource/purchase_datasource.dart';
 import 'package:quickdrop/src/data/datasource/purchase_detail_datasource.dart';
 import 'package:quickdrop/src/features/auth/data/repository/auth_repository_impl.dart';
 import 'package:quickdrop/src/features/favorites/data/repository/favorite_repository_impl.dart';
@@ -15,7 +15,7 @@ import 'package:quickdrop/src/features/home/data/repository/home_data_repository
 import 'package:quickdrop/src/data/repository/my_purchases_repository_impl.dart';
 import 'package:quickdrop/src/features/productDetail/data/repository/product_detail_repository_impl.dart';
 import 'package:quickdrop/src/data/repository/purchase_detail_repository_impl.dart';
-import 'package:quickdrop/src/data/repository/purchase_repository_impl.dart';
+import 'package:quickdrop/src/features/pruchase/data/repository/purchase_repository_impl.dart';
 import 'package:quickdrop/src/features/auth/domain/repository/auth_repository.dart';
 import 'package:quickdrop/src/features/favorites/domain/repository/favorite_repository.dart';
 import 'package:quickdrop/src/domain/repository/financial_information_repository.dart';
@@ -23,7 +23,7 @@ import 'package:quickdrop/src/features/home/domain/repository/home_data_reposito
 import 'package:quickdrop/src/domain/repository/my_purchases_repository.dart';
 import 'package:quickdrop/src/features/productDetail/domain/repository/product_detail_repository.dart';
 import 'package:quickdrop/src/domain/repository/purchase_detail_repository.dart';
-import 'package:quickdrop/src/domain/repository/purchase_repository.dart';
+import 'package:quickdrop/src/features/pruchase/domain/repository/purchase_repository.dart';
 import 'package:quickdrop/src/features/auth/domain/usecase/auth_usecase.dart';
 import 'package:quickdrop/src/features/favorites/domain/usecase/favorite_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/financial_information_usecase.dart';
@@ -31,7 +31,7 @@ import 'package:quickdrop/src/features/home/domain/usecase/home_data_usecase.dar
 import 'package:quickdrop/src/domain/usecase/my_purchases_usecase.dart';
 import 'package:quickdrop/src/features/productDetail/domain/usecase/produc_detail_usecase.dart';
 import 'package:quickdrop/src/domain/usecase/purchase_detail_usecase.dart';
-import 'package:quickdrop/src/domain/usecase/purshase_usecase.dart';
+import 'package:quickdrop/src/features/pruchase/domain/usecase/purshase_usecase.dart';
 import 'package:quickdrop/src/features/app/cubit/app_cubit.dart';
 import 'package:quickdrop/src/features/favorites/presentation/cubit/favorites_cubit.dart';
 import 'package:quickdrop/src/presentation/financial_information/cubit/financial_information_cubit.dart';
